version 6.0 , first formal release
版本代号：Thank You

> "Thank You" is a song written and performed by English singer-songwriter
Dido. The song, released on her 1999 debut studio album, No Angel, made
its first appearance in 1998, on the soundtrack of Sliding Doors,
although it was not released as a single until late 2000. It was later
featured on Dido's debut album, No Angel, becoming the album's biggest
hit. The song's first verse is notably used as the chorus in the song
"Stan" by Eminem.[1]

<https://en.wikipedia.org/wiki/Thank_You_(Dido_song)>