player: the player skipped a song unexpectedly (#948)
Before, feeluown stops the MPV by playing a `''` file. The action causes the MPV to emit an EndFile event in the new version of libmpv. In the old version of libmpv, it does not emit any event. This change causes the bug. 
```
[v] bdmv/bluray: Opening
[v] file: Opening
[error] file: Cannot open file '': No such file or directory
[error] stream: Failed to open .
[v] cplayer: Opening failed or was aborted:
[v] cplayer: finished playback, loading failed (reason 4)
[v] stream_callback: Opening http://xxxx
```

Now, call `stop` method directly to stop the MPV instead of `play('')`